youtube-shorts-generator/requirements.txt
fastapi
uvicorn
moviepy
gTTS
transformers
nltk
requests
beautifulsoup4
opencv-python
numpy
pandas
python-dotenv
